move MLP definition
