+++ ambersim/learning/architectures.py
@@ -1,0 +1,32 @@
+from typing import Sequence
+
+import flax.linen as nn
+import jax.numpy as jnp
+
+
+class MLP(nn.Module):
+    """A simple pickle-able multi-layer perceptron.
+
+    Args:
+        layer_sizes: Sizes of all hidden layers and the output layer.
+        activate_final: Whether to apply an activation function to the output.
+        bias: Whether to use a bias in the linear layers.
+    """
+
+    layer_sizes: Sequence[int]
+    activate_final: bool = False
+    bias: bool = True
+
+    @nn.compact
+    def __call__(self, x: jnp.ndarray) -> jnp.ndarray:
+        """Forward pass through the network."""
+        # TODO(vincekurtz): consider using jax control flow here
+        for i, layer_size in enumerate(self.layer_sizes):
+            x = nn.Dense(
+                layer_size,
+                use_bias=self.bias,
+                name=f"dense_{i}",
+            )(x)
+            if i != len(self.layer_sizes) - 1 or self.activate_final:
+                x = nn.relu(x)
+        return x
